pyproject.toml
[project]
name = "fal-api-mcp-server"
version = "0.1.0"
description = "A MCP server project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fal-client>=0.5.9",
    "httpx>=0.28.1",
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3"
]

[[project.authors]]
name = "taimo3810"
email = "132860814+taimo3810@users.noreply.github.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
fal-api-mcp-server = "fal_api_mcp_server:main"
